+++ src/portals/hope/features/search/SearchInput.tsx
@@ -1,11 +1,6 @@
-import React, { useEffect, useState } from 'react'
+import React, { useState } from 'react'
 import { SearchIcon } from 'portals/hope/features/members-search/styles'
 import { isPressing, Keys } from '@hedvig-ui/hooks/keyboard/use-key-is-pressed'
-import { ArrayElement } from '@hedvig-ui/utils/array-element'
-import {
-  useSearch,
-  UseSearchResult,
-} from 'portals/hope/common/hooks/use-search'
 import styled from '@emotion/styled'
 import { Input } from '@hedvig-ui'
 
@@ -24,39 +19,13 @@ const SuggestionText = styled.div`
   margin-bottom: -2.86rem;
 `
 
-const useAutoComplete = (
-  query: string,
-): ArrayElement<UseSearchResult['hits']> | null => {
-  const { hits, search } = useSearch(query, {
-    type: 'FULL_NAME',
-  })
-
-  useEffect(() => search(), [query])
-
-  return (
-    hits.find((hit) =>
-      `${hit.firstName} ${hit.lastName}`
-        .toLowerCase()
-        .startsWith(query.toLowerCase()),
-    ) ?? null
-  )
-}
-
 export const SearchInput: React.FC<{
+  onChange: (query: string) => void
   onSearch: (query: string) => void
   loading?: boolean
-}> = ({ onSearch, loading }) => {
+  suggestion?: string
+}> = ({ onSearch, onChange, loading, suggestion = '' }) => {
   const [value, setValue] = useState('')
-  const suggestion = useAutoComplete(value)
-
-  const suggestionString = () => {
-    if (!value) return ''
-    if (!suggestion?.firstName || !suggestion?.lastName) return ''
-
-    const completeString = `${suggestion.firstName} ${suggestion.lastName}`
-
-    return value + completeString.substring(value.length)
-  }
 
   return (
     <form
@@ -69,6 +38,7 @@ export const SearchInput: React.FC<{
         <StyledInput
           onChange={(e) => {
             setValue(e.currentTarget.value)
+            onChange(e.currentTarget.value)
           }}
           value={value}
           size="large"
@@ -80,15 +50,15 @@ export const SearchInput: React.FC<{
           onKeyDown={(e) => {
             if (
               isPressing(e, Keys.Right) &&
-              suggestionString() &&
-              suggestionString() !== value
+              suggestion &&
+              suggestion !== value
             ) {
-              setValue(suggestionString)
+              setValue(suggestion)
             }
           }}
         />
 
-        <SuggestionText>{suggestionString() || '\u00a0'}</SuggestionText>
+        <SuggestionText>{suggestion || '\u00a0'}</SuggestionText>
       </div>{' '}
     </form>
   )
