+++ src/portals/hope/pages/search/SearchPage.tsx
@@ -1,336 +1,9 @@
 import { Page } from 'portals/hope/pages/routes'
-import {
-  Button,
-  Flex,
-  Loadable,
-  Placeholder,
-  Popover,
-  Spacing,
-  Table,
-  TableBody,
-  TableColumn,
-  TableHeader,
-  TableHeaderColumn,
-  TableRow,
-} from '@hedvig-ui'
-import styled from '@emotion/styled'
-import React, { useEffect, useState } from 'react'
-import { useHistory } from 'react-router'
-import { useSearch } from '../../common/hooks/use-search'
-import parse from 'html-react-parser'
-import chroma from 'chroma-js'
-import {
-  Contract,
-  ContractStatus,
-  SearchQuery,
-  useMemberSearchResultQuery,
-} from 'types/generated/graphql'
-import { ArrayElement } from '@hedvig-ui/utils/array-element'
-import { MemberAge } from 'portals/hope/features/member/utils'
-import formatDate from 'date-fns/format'
-import { parseISO } from 'date-fns'
-import gql from 'graphql-tag'
-import {
-  getFirstMasterInception,
-  getLastTerminationDate,
-} from 'portals/hope/features/member/tabs/contracts-tab/utils'
-import { SearchInput } from 'portals/hope/features/search/SearchInput'
-
-type CircleVariation =
-  | 'success'
-  | 'warning'
-  | 'danger'
-  | 'accent'
-  | 'placeholderColor'
-
-const ContractCountWrapper = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-`
-
-const ContractCountNumber = styled.div<{ variation?: CircleVariation }>`
-  color: white;
-  background-color: ${({ theme, variation = 'accent' }) => theme[variation]};
-  padding: 0.2em 0.6em;
-  border-radius: 6px;
-  font-size: 0.8em;
-  min-width: 2em;
-  text-align: center;
-`
-
-const ContractCountLabel = styled.div`
-  font-size: 0.7em;
-  color: ${({ theme }) => theme.semiStrongForeground};
-  margin-top: 0.3em;
-`
-
-const TableColumnSubtext = styled.span`
-  font-size: 0.8em;
-  color: ${({ theme }) => theme.semiStrongForeground};
-`
-
-const MemberAgeWrapper = styled.span`
-  font-size: 0.7em;
-  margin-left: 0.7em;
-  margin-right: -0.7em;
-  color: ${({ theme }) => theme.semiStrongForeground};
-`
-
-const SearchHitTag = styled.div`
-  background-color: ${({ theme }) =>
-    chroma(theme.semiStrongForeground).brighten(2.25).alpha(0.5).hex()};
-  color: ${({ theme }) => theme.semiStrongForeground};
-
-  padding: 0.15rem 0.3rem;
-  font-size: 0.8rem;
-  margin-right: 0.5rem;
-  border-radius: 0.25rem;
-`
-
-const SearchHitExplanation = styled.div`
-  color: ${({ theme }) => theme.semiStrongForeground};
-  font-size: 0.8rem;
-  margin-right: 0.5rem;
-`
-
-const convertTagText = (text: string) => {
-  const result = text.replaceAll(/([A-Z])/g, ' $1')
-  return result.charAt(0).toUpperCase() + result.slice(1)
-}
-
-gql`
-  query MemberSearchResult($memberId: ID!) {
-    member(id: $memberId) {
-      memberId
-      birthDate
-      signedOn
-      contracts {
-        id
-        masterInception
-        terminationDate
-        status
-      }
-    }
-  }
-`
-
-const countContractsByStatus = (contractStatuses: ContractStatus[]) =>
-  contractStatuses.reduce<Record<string, number>>((acc, status) => {
-    const groupedStatus = [
-      ContractStatus.Pending,
-      ContractStatus.Terminated,
-      ContractStatus.ActiveInFuture,
-    ].includes(status)
-      ? status
-      : ContractStatus.Active
-    return {
-      ...acc,
-      [groupedStatus]: (acc[groupedStatus] || 0) + 1,
-    }
-  }, {})
-
-const SearchResult: React.FC<{
-  result: ArrayElement<SearchQuery['search']>
-}> = ({ result }) => {
-  const history = useHistory()
-  const { data, loading } = useMemberSearchResultQuery({
-    variables: { memberId: result.memberId ?? '' },
-  })
-
-  const contracts = (data?.member?.contracts ?? []) as Array<Contract>
-
-  const {
-    [ContractStatus.ActiveInFuture]: activeInFutureContracts = 0,
-    [ContractStatus.Active]: activeContracts = 0,
-    [ContractStatus.Pending]: pendingContracts = 0,
-    [ContractStatus.Terminated]: terminatedContracts = 0,
-  } = countContractsByStatus(contracts.map((contract) => contract.status))
-
-  return (
-    <TableRow
-      tabIndex={0}
-      onClick={() => history.push(`/members/${result.memberId}/contracts`)}
-    >
-      <TableColumn style={{ verticalAlign: 'top' }}>
-        <Flex direction="column">
-          {result.firstName && result.lastName ? (
-            `${result.firstName} ${result.lastName}`
-          ) : (
-            <Placeholder>Not available</Placeholder>
-          )}
-          <Flex>
-            <TableColumnSubtext>{result.memberId}</TableColumnSubtext>
-            <MemberAgeWrapper>
-              <MemberAge birthDateString={data?.member?.birthDate} />
-            </MemberAgeWrapper>
-          </Flex>
-          {result.highlights.filter(
-            (highlight) => !highlight.field.includes('keyword'),
-          ).length !== 0 && (
-            <Flex align="center" style={{ marginTop: '1rem' }}>
-              <SearchHitExplanation>Search hit in </SearchHitExplanation>
-              {result.highlights
-                .filter((highlight) => !highlight.field.includes('keyword'))
-                .map((highlight) => (
-                  <SearchHitTag key={highlight.field}>
-                    <Popover
-                      style={{
-                        maxWidth: '60rem',
-                        minWidth: '15rem',
-                        overflowWrap: 'break-word',
-                      }}
-                      contents={parse(
-                        [...new Set(highlight.values)]
-                          .reduce<string>(
-                            (acc, value) => acc + value + '<br/>',
-                            '',
-                          )
-                          .replaceAll('<em>', '<b>')
-                          .replaceAll('</em>', '</b>'),
-                      )}
-                    >
-                      {convertTagText(highlight.field).replaceAll('.', ', ')}
-                    </Popover>
-                  </SearchHitTag>
-                ))}
-            </Flex>
-          )}
-        </Flex>
-      </TableColumn>
-
-      <TableColumn style={{ verticalAlign: 'top' }}>
-        {data?.member?.signedOn ? (
-          <Flex direction="column">
-            {formatDate(parseISO(data?.member?.signedOn), 'dd MMMM, yyyy')}
-            <TableColumnSubtext>
-              {formatDate(parseISO(data?.member?.signedOn), 'HH:mm')}
-            </TableColumnSubtext>
-          </Flex>
-        ) : (
-          <Placeholder>Not available</Placeholder>
-        )}
-      </TableColumn>
-
-      <TableColumn style={{ verticalAlign: 'top' }}>
-        <Loadable loading={loading}>
-          {getFirstMasterInception(contracts) ?? (
-            <Placeholder>Not specified</Placeholder>
-          )}
-        </Loadable>
-      </TableColumn>
-      <TableColumn style={{ verticalAlign: 'top' }}>
-        <Loadable loading={loading}>
-          {getLastTerminationDate(contracts) ?? (
-            <Placeholder>Not specified</Placeholder>
-          )}
-        </Loadable>
-      </TableColumn>
-
-      <TableColumn style={{ verticalAlign: 'top' }}>
-        <Flex justify="space-between">
-          <Loadable loading={loading}>
-            <ContractCountWrapper>
-              <ContractCountNumber
-                variation={pendingContracts ? 'warning' : 'placeholderColor'}
-              >
-                {pendingContracts}
-              </ContractCountNumber>
-              <ContractCountLabel>Pending</ContractCountLabel>
-            </ContractCountWrapper>
-          </Loadable>
-          <Loadable loading={loading}>
-            <ContractCountWrapper>
-              <ContractCountNumber
-                variation={
-                  activeInFutureContracts ? 'accent' : 'placeholderColor'
-                }
-              >
-                {activeInFutureContracts}
-              </ContractCountNumber>
-              <ContractCountLabel>Active in future</ContractCountLabel>
-            </ContractCountWrapper>
-          </Loadable>
-          <Loadable loading={loading}>
-            <ContractCountWrapper>
-              <ContractCountNumber
-                variation={activeContracts ? 'success' : 'placeholderColor'}
-              >
-                {activeContracts}
-              </ContractCountNumber>
-              <ContractCountLabel>Active</ContractCountLabel>
-            </ContractCountWrapper>
-          </Loadable>
-          <Loadable loading={loading}>
-            <ContractCountWrapper>
-              <ContractCountNumber
-                variation={terminatedContracts ? 'danger' : 'placeholderColor'}
-              >
-                {terminatedContracts}
-              </ContractCountNumber>
-              <ContractCountLabel>Terminated</ContractCountLabel>
-            </ContractCountWrapper>
-          </Loadable>
-        </Flex>
-      </TableColumn>
-    </TableRow>
-  )
-}
+import React from 'react'
+import { MemberSearch } from 'portals/hope/features/search/members/MemberSearch'
 
 const SearchPage: Page = () => {
-  const [query, setQuery] = useState('')
-  const { hits, loading, search, fetchMore } = useSearch(query)
-
-  useEffect(() => {
-    if (query) {
-      search()
-    }
-  }, [query])
-
-  return (
-    <>
-      <SearchInput
-        onSearch={(value) => {
-          setQuery(value)
-        }}
-        loading={loading}
-      />
-
-      <Spacing top="large" />
-
-      {hits.length !== 0 && (
-        <Table>
-          <TableHeader>
-            <TableHeaderColumn>Member</TableHeaderColumn>
-            <TableHeaderColumn>Signed Up</TableHeaderColumn>
-            <TableHeaderColumn>First Master Inception</TableHeaderColumn>
-            <TableHeaderColumn>Last Termination Date</TableHeaderColumn>
-            <TableHeaderColumn>Contracts</TableHeaderColumn>
-          </TableHeader>
-          <TableBody>
-            {hits.map((member) => (
-              <React.Fragment key={member.memberId}>
-                <SearchResult result={member} />
-              </React.Fragment>
-            ))}
-          </TableBody>
-        </Table>
-      )}
-      <Spacing top="medium" />
-      <Flex justify="center">
-        {hits.length !== 0 && hits.length >= 20 && (
-          <Button
-            variant="tertiary"
-            onClick={() => {
-              fetchMore()
-            }}
-          >
-            Show more
-          </Button>
-        )}
-      </Flex>
-    </>
-  )
+  return <MemberSearch />
 }
 
 export default SearchPage
